refactor: data management refinements from cross-testing in astro
