+++ include/OpenSpaceToolkit/Physics/Coordinate/Frame/Providers/IERS/Manager.hpp
@@ -19,7 +19,7 @@
 
 #define OSTK_PHYSICS_COORDINATE_FRAME_PROVIDERS_IERS_MANAGER_MODE Manager::Mode::Automatic
 #define OSTK_PHYSICS_COORDINATE_FRAME_PROVIDERS_IERS_MANAGER_LOCAL_REPOSITORY \
-    "./.open-space-toolkit/physics/coordinate/frame/providers/iers"
+    "./.open-space-toolkit/physics/data/coordinate/frame/providers/iers"
 #define OSTK_PHYSICS_COORDINATE_FRAME_PROVIDERS_IERS_MANAGER_LOCAL_REPOSITORY_LOCK_TIMEOUT 60
 
 namespace ostk
